factor a lot of extension functions to new file
